Move SawtoothConnection under sawtooth module
Move the SawtoothConnection under the sawtooth module, to cleanly
partition the sawtooth code. It is in the `sawtooth::connection` module
specifically.
