+++ daemon/src/sawtooth/mod.rs
@@ -14,3 +14,5 @@
  * limitations under the License.
  * -----------------------------------------------------------------------------
  */
+
+pub mod connection;
